Move d3, tsk, and color utils into one folder
