+++ app/assets/javascripts/components/line_plot_attribute.js.jsx
@@ -1,4 +1,4 @@
-import { autoColors } from '../colors'
+import { autoColors } from '../utils/colors'
 
 var LinePlotAttribute = React.createClass({
   render: function() {
