+++ app/assets/javascripts/components/plots/box_plot.jsx
@@ -1,4 +1,4 @@
-import { createScale } from './utils/d3_scale'
+import { createScale } from '../../utils/d3_scale'
 
 var BoxPlot = React.createClass({
   render: function() {
