+++ app/assets/javascripts/utils/colors.js
